:recycle: Refactor the app to adopt BEM approach (6)
