+++ src/app/classic.blocks/MousePositionPresentation.scss
@@ -1,0 +1,4 @@
+#MousePositionPresentation {
+  margin-top: 16px !important;
+  user-select: none;
+}
