+++ src/app/classic.blocks/_index.scss
@@ -83,14 +83,6 @@ $middle-section-height-svh: calc(100svh - $pg - $pg);
   overflow-y: auto;
   height: calc(100% - $hpg - $hpg + $hpg);
   position: relative;
-
-  #MousePositionPresentation {
-    position: absolute;
-    left: 0;
-    right: 0;
-    top: calc(100% + 16px);
-    user-select: none;
-  }
 }
 
 #LeftSection::-webkit-scrollbar,
@@ -118,7 +110,10 @@ $middle-section-height-svh: calc(100svh - $pg - $pg);
   margin-top: $hpg;
 }
 
+// Import all blocks scss file that without tsx implementation
+
 @import "./ControlAccordion";
 @import "./MenuAccordion";
+@import "./MousePositionPresentation";
 @import "./SpeedCanvasElement";
 @import "./PathTreeAccordion";
